Add an object for finding Crosswalk pieces
The intention is to make it possible to define the "pieces"
inside a Crosswalk Android tarball, in a declarative way. Then
if pieces move around, new jar files added etc.,
we can still find the pieces and assign them to the correct
properties in the Env (environment) object so that a build is
still possible.