Env should sort results returned by Finder
The results of a file search (e.g. for aapt) should be returned
as plain arrays (Android SDK agnostic) to Env; then Env should
take responsibility for choosing the "latest" matching file.

This is done by passing a filter function to findFiles(), which
optionally selects a single item from the array of files
returned (either by glob or by guessing directories). The
single returned path can then be returned as the location of the
desired "piece".

findDirectories() also now returns multiple matching directories,
and the last of them is selected as the desired "piece".

This allows all code which is Android-specific to stay in Env and
not have to go into Finder.