Only use arch in suffix when embedded is true
This also moves the embedded property to Env (from App).

I also changed the signature for the Locations() constructor
to Locations(app, env, outDir), to make it easier to get
the properties out of the objects (you have access to all properties
now, rather than having to explicitly pass them to Locations).

Fixes #39